+++ src/main/java/com/example/MercadosoBack/services/UserService.java
@@ -19,7 +19,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public UserModel getUserById(Integer id) {
+    public UserModel getUserById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
 
@@ -27,7 +27,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
